devops/metrics/src/webhook_catchers/netlify/requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
idna==3.6
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
urllib3==2.1.0
